Commenter: Move into commenter package
