+++ resources/META-INF/plugin.xml
@@ -49,8 +49,7 @@
 
         <!-- Commenter -->
 
-        <lang.commenter language="RUST" implementationClass="org.rust.lang.RustCommenter"/>
-
+        <lang.commenter language="RUST" implementationClass="org.rust.lang.commenter.RustCommenter"/>
         <!-- Matchers -->
 
         <lang.braceMatcher language="RUST" implementationClass="org.rust.lang.highlight.RustBraceMatcher"/>
