+++ test/org/rust/lang/commenter/RustCommenterTest.kt
@@ -1,6 +1,7 @@
-package org.rust.lang
+package org.rust.lang.commenter
 
 import com.intellij.openapi.actionSystem.IdeActions
+import org.rust.lang.RustTestCase
 
 /**
  * @see RustCommenter
